create evm base class
