+++ packages/bitcore-node/src/modules/ethereum/api/gnosis.ts
@@ -1,16 +1,17 @@
 import { Readable } from 'stream';
 import { Transaction } from 'web3-eth';
 import { AbiItem } from 'web3-utils';
+import { MultisigAbi } from '../../../providers/chain-state/evm/abi/multisig';
+import { MultisigRelatedFilterTransform } from '../../../providers/chain-state/evm/api/multisigTransform';
+import { PopulateReceiptTransform } from '../../../providers/chain-state/evm/api/populateReceiptTransform';
+import { EVMListTransactionsStream } from '../../../providers/chain-state/evm/api/transform';
+import { EVMBlockStorage } from '../../../providers/chain-state/evm/models/block';
+import { EVMTransactionStorage } from '../../../providers/chain-state/evm/models/transaction';
+import { EventLog } from '../../../providers/chain-state/evm/types';
 import { Config } from '../../../services/config';
-import { IEthNetworkConfig } from '../../../types/Config';
+import { IEVMNetworkConfig } from '../../../types/Config';
 import { StreamWalletTransactionsParams } from '../../../types/namespaces/ChainStateProvider';
-import { MultisigAbi } from '../abi/multisig';
-import { EthBlockStorage } from '../models/block';
-import { EthTransactionStorage } from '../models/transaction';
-import { ETH, EventLog } from './csp';
-import { EthMultisigRelatedFilterTransform } from './ethMultisigTransform';
-import { PopulateReceiptTransform } from './populateReceiptTransform';
-import { EthListTransactionsStream } from './transform';
+import { ETH } from './csp';
 
 interface MULTISIGInstantiation
   extends EventLog<{
@@ -42,10 +43,10 @@ export class GnosisApi {
     txId: string
   ): Promise<Partial<Transaction>[]> {
     const { web3 } = await ETH.getWeb3(network);
-    const networkConfig: IEthNetworkConfig = Config.chainConfig({ chain: 'ETH', network });
+    const networkConfig: IEVMNetworkConfig = Config.chainConfig({ chain: 'ETH', network });
     const { gnosisFactory = this.gnosisFactories[network] } = networkConfig;
     let query = { chain: 'ETH', network, txid: txId };
-    const found = await EthTransactionStorage.collection.findOne(query);
+    const found = await EVMTransactionStorage.collection.findOne(query);
     const blockHeight = found && found.blockHeight ? found.blockHeight : null;
     if (!blockHeight || blockHeight < 0) return Promise.resolve([]);
     const contract = await this.multisigFor(network, gnosisFactory);
@@ -78,7 +79,7 @@ export class GnosisApi {
   async getMultisigTxpsInfo(network: string, multisigContractAddress: string): Promise<Partial<Transaction>[]> {
     const contract = await this.multisigFor(network, multisigContractAddress);
     const time = Math.floor(Date.now()) - this.ETH_MULTISIG_TX_PROPOSAL_EXPIRE_TIME;
-    const [block] = await EthBlockStorage.collection
+    const [block] = await EVMBlockStorage.collection
       .find({
         chain: 'ETH',
         network,
@@ -175,20 +176,16 @@ export class GnosisApi {
     }
 
     let transactionStream = new Readable({ objectMode: true });
-    const ethTransactionTransform = new EthListTransactionsStream([multisigContractAddress, args.tokenAddress]);
+    const ethTransactionTransform = new EVMListTransactionsStream([multisigContractAddress, args.tokenAddress]);
     const populateReceipt = new PopulateReceiptTransform();
 
-    transactionStream = EthTransactionStorage.collection
+    transactionStream = EVMTransactionStorage.collection
       .find(query)
       .sort({ blockTimeNormalized: 1 })
       .addCursorFlag('noCursorTimeout', true);
 
     if (multisigContractAddress) {
-      const ethMultisigTransform = new EthMultisigRelatedFilterTransform(
-        web3,
-        multisigContractAddress,
-        args.tokenAddress
-      );
+      const ethMultisigTransform = new MultisigRelatedFilterTransform(web3, multisigContractAddress, args.tokenAddress);
       transactionStream = transactionStream.pipe(ethMultisigTransform);
     }
