+++ packages/bitcore-node/src/modules/ethereum/index.ts
@@ -1,15 +1,15 @@
 import { BaseModule } from '..';
+import { EVMVerificationPeer } from '../,,/../../providers/chain-state/evm/p2p/EVMVerificationPeer';
 import { ETHStateProvider } from './api/csp';
 import { EthRoutes } from './api/eth-routes';
-import { EthVerificationPeer } from './p2p/EthVerificationPeer';
 import { EthP2pWorker } from './p2p/p2p';
 
 export default class ETHModule extends BaseModule {
   constructor(services: BaseModule['bitcoreServices']) {
     super(services);
     services.P2P.register('ETH', EthP2pWorker);
-    services.CSP.registerService('ETH', new ETHStateProvider());
+    services.CSP.registerService('ETH', new ETHStateProvider('ETH'));
     services.Api.app.use(EthRoutes);
-    services.Verification.register('ETH', EthVerificationPeer);
+    services.Verification.register('ETH', EVMVerificationPeer);
   }
 }
