+++ packages/bitcore-node/src/providers/chain-state/evm/abi/erc20.ts
