+++ packages/bitcore-node/src/providers/chain-state/evm/abi/erc721.ts
