+++ packages/bitcore-node/src/providers/chain-state/evm/abi/multisig.ts
