+++ packages/bitcore-node/src/providers/chain-state/evm/p2p/rpcs/gethRpc.ts
@@ -1,6 +1,6 @@
 import Web3 from 'web3';
-import { EthTransactionStorage } from '../../models/transaction';
-import { GethBlock, IAbiDecodedData, IEthBlock, IEthTransaction } from '../../types';
+import { EVMTransactionStorage } from '../../models/transaction';
+import { GethBlock, IAbiDecodedData, IEVMBlock, IEVMTransaction } from '../../types';
 import { Callback, IJsonRpcRequest, IJsonRpcResponse, IRpc } from './index';
 
 interface IGethTxTraceResponse {
@@ -49,7 +49,7 @@ export class GethRPC implements IRpc {
   }
 
   public async getTransactionsFromBlock(blockNumber: number): Promise<IGethTxTrace[]> {
-    const txs = (await this.traceBlock(blockNumber)) || [];
+    const txs = (await this.traceBlock(blockNumber)).filter(tx => tx.result) || [];
     return txs.map(tx => this.transactionFromGethTrace(tx));
   }
 
@@ -65,15 +65,15 @@ export class GethRPC implements IRpc {
 
   private transactionFromGethTrace(tx: IGethTxTraceResponse) {
     const convertedTx = tx.result;
-    convertedTx.abiType = EthTransactionStorage.abiDecode(tx.result.input);
+    convertedTx.abiType = EVMTransactionStorage.abiDecode(tx.result.input);
 
     for (let call of convertedTx.calls || []) {
-      call.abiType = EthTransactionStorage.abiDecode(tx.result.input);
+      call.abiType = EVMTransactionStorage.abiDecode(tx.result.input);
     }
     return convertedTx;
   }
 
-  public reconcileTraces(block: IEthBlock, transactions: IEthTransaction[], traces: IGethTxTrace[]) {
+  public reconcileTraces(block: IEVMBlock, transactions: IEVMTransaction[], traces: IGethTxTrace[]) {
     // TODO calculate total block reward including fees
     block;
 
@@ -92,7 +92,7 @@ export class GethRPC implements IRpc {
     const calls = trace.calls;
 
     delete trace.calls;
-    trace.abiType = trace.input ? EthTransactionStorage.abiDecode(trace.input) : undefined;
+    trace.abiType = trace.input ? EVMTransactionStorage.abiDecode(trace.input) : undefined;
     (trace as IGethTxTraceFlat).depth = depth;
     retval.push(trace as IGethTxTraceFlat);
