+++ packages/bitcore-node/test/integration/ethereum/p2p.spec.ts
@@ -4,10 +4,10 @@ import { Web3 } from 'crypto-wallet-core';
 import sinon from 'sinon';
 import config from '../../../src/config';
 import { CacheStorage } from '../../../src/models/cache';
-import { EthBlockStorage } from '../../../src/modules/ethereum/models/block';
+import { EVMBlockStorage } from '../../../src/providers/chain-state/evm/models/block';
 import { EthP2pWorker } from '../../../src/modules/ethereum/p2p/p2p';
 import { Api } from '../../../src/services/api';
-import { IEthNetworkConfig } from '../../../src/types/Config';
+import { IEVMNetworkConfig } from '../../../src/types/Config';
 import { wait } from '../../../src/utils/wait';
 import { resetDatabase } from '../../helpers';
 import { intAfterHelper, intBeforeHelper } from '../../helpers/integration';
@@ -15,7 +15,7 @@ import { intAfterHelper, intBeforeHelper } from '../../helpers/integration';
 const { StreamUtil } = BitcoreClient;
 const chain = 'ETH';
 const network = 'regtest';
-const chainConfig = config.chains[chain][network] as IEthNetworkConfig;
+const chainConfig = config.chains[chain][network] as IEVMNetworkConfig;
 const name = 'EthereumWallet-Ci';
 const baseUrl = 'http://localhost:3000/api';
 const password = '';
@@ -189,7 +189,7 @@ describe('Ethereum', function() {
     await done;
     await worker.stop();
 
-    const dbBlocks = await EthBlockStorage.collection.count({ chain, network });
+    const dbBlocks = await EVMBlockStorage.collection.count({ chain, network });
     expect(dbBlocks).to.be.gt(0);
     await wallet.lock();
   });
