+++ packages/bitcore-node/test/integration/matic/p2p.spec.ts
@@ -1,23 +1,25 @@
 import * as BitcoreClient from 'bitcore-client';
 import { expect } from 'chai';
 import config from '../../../src/config';
+import { Web3 } from 'crypto-wallet-core';
 import { CacheStorage } from '../../../src/models/cache';
-import { EthBlockStorage } from '../../../src/modules/ethereum/models/block';
-import { EthP2pWorker } from '../../../src/modules/ethereum/p2p/p2p';
+import { EVMBlockStorage } from '../../../src/providers/chain-state/evm/models/block';
+import { EVMP2pWorker } from '../../../src/providers/chain-state/evm/p2p/p2p';
 import { Api } from '../../../src/services/api';
-import { IEthNetworkConfig } from '../../../src/types/Config';
+import { IEVMNetworkConfig } from '../../../src/types/Config';
 import { wait } from '../../../src/utils/wait';
 import { resetDatabase } from '../../helpers';
 import { intAfterHelper, intBeforeHelper } from '../../helpers/integration';
+import sinon from 'sinon';
 
 const { StreamUtil } = BitcoreClient;
 const chain = 'MATIC';
 const network = 'regtest';
-const chainConfig = config.chains[chain][network] as IEthNetworkConfig;
+const chainConfig = config.chains[chain][network] as IEVMNetworkConfig;
 const name = 'PolygonWallet-Ci';
 const baseUrl = 'http://localhost:3000/api';
 const password = '';
-const phrase = 'kiss talent nerve fossil equip fault exile execute train wrist misery diet';
+const phrase = 'glimpse mystery poverty onion muffin twist live kidney unhappy sort frame muffin';
 const accounts = { geth: '0xeC12CD1Ab86F83C1B26C5caa38126Bc4299b6CBa' };
 const privKeys = { geth: '0xf9ad2207e910cd649c9a32063dea3656380c32fa07d6bb9be853687ca585a015' };
 
@@ -76,62 +78,62 @@ describe('Polygon', function() {
     const addresses = await wallet.getAddresses();
     expect(addresses).to.exist;
     expect(addresses.length).to.eq(1);
-    expect(addresses[0].toLowerCase()).to.equal('0xaf5b457a1687919bfb078f12947381559f79e36a');
+    expect(addresses[0].toLowerCase()).to.equal('0xb875c670b079cce8f6ab4f013bd471359c877ac0');
   });
 
-  // it('should be able to get block events from geth', async () => {
-  //   const gethOnlyConfig = { ...chainConfig, provider: chainConfig.providers![0] };
-  //   const { protocol, host, port } = gethOnlyConfig.provider;
-  //   const getWeb3Stub = sinon.stub(EthP2pWorker.prototype, 'getWeb3').resolves({ web3: new Web3(`${protocol}://${host}:${port}`) });
+  it('should be able to get block events from geth', async () => {
+    const gethOnlyConfig = { ...chainConfig, provider: chainConfig.providers![0] };
+    const { protocol, host, port } = gethOnlyConfig.provider;
+    const getWeb3Stub = sinon.stub(EVMP2pWorker.prototype, 'getWeb3').resolves({ web3: new Web3(`${protocol}://${host}:${port}`) });
 
-  //   const wallet = await getWallet();
-  //   const addresses = await wallet.getAddresses();
+    const wallet = await getWallet();
+    const addresses = await wallet.getAddresses();
 
-  //   const worker = new EthP2pWorker({ chain, network, chainConfig: gethOnlyConfig });
-  //   await worker.setupListeners();
-  //   await worker.connect();
-  //   const sawBlock = new Promise(resolve => worker.events.on('block', resolve));
+    const worker = new EVMP2pWorker({ chain, network, chainConfig: gethOnlyConfig });
+    await worker.setupListeners();
+    await worker.connect();
+    const sawBlock = new Promise(resolve => worker.events.on('block', resolve));
 
-  //   const { web3 } = await worker.getWeb3();
-  //   await sendTransaction('geth', addresses[0], web3.utils.toWei('.01', 'ether'), web3, wallet);
-  //   await sawBlock;
-  //   await worker.disconnect();
-  //   await worker.stop();
-  //   getWeb3Stub.restore();
-  // });
+    const { web3 } = await worker.getWeb3();
+    await sendTransaction('geth', addresses[0], web3.utils.toWei('.01', 'ether'), web3, wallet);
+    await sawBlock;
+    await worker.disconnect();
+    await worker.stop();
+    getWeb3Stub.restore();
+  });
 
   it('should be able to get the balance for the address', async () => {
     const wallet = await getWallet();
     const balance = await wallet.getBalance();
-    expect(balance.confirmed).to.be.eq(0);
+    expect(balance.confirmed).to.be.gt(0);
 
-    const key = 'getBalanceForAddress-MATIC-regtest-0xaf5b457a1687919bfb078f12947381559f79e36a';
+    const key = 'getBalanceForAddress-MATIC-regtest-0xb875c670b079cce8f6ab4f013bd471359c877ac0';
     const cached = await CacheStorage.collection.findOne({ key });
     expect(cached).to.exist;
     expect(cached!.value).to.deep.eq(balance);
     await wallet.lock();
   });
 
-  // it('should update after a send', async () => {
-  //   const wallet = await getWallet();
-  //   const addresses = await wallet.getAddresses();
-  //   const beforeBalance = await wallet.getBalance();
+  it('should update after a send', async () => {
+    const wallet = await getWallet();
+    const addresses = await wallet.getAddresses();
+    const beforeBalance = await wallet.getBalance();
 
-  //   const worker = new EthP2pWorker({ chain, network, chainConfig });
-  //   await worker.setupListeners();
-  //   await worker.connect();
-  //   const sawBlock = new Promise(resolve => worker.events.on('block', resolve));
+    const worker = new EVMP2pWorker({ chain, network, chainConfig });
+    await worker.setupListeners();
+    await worker.connect();
+    const sawBlock = new Promise(resolve => worker.events.on('block', resolve));
 
-  //   const { web3 } = await worker.getWeb3();
-  //   await sendTransaction('geth', addresses[0], web3.utils.toWei('.01', 'ether'), web3, wallet);
-  //   await sawBlock;
-  //   await worker.disconnect();
-  //   await worker.stop();
-  //   const afterBalance = await wallet.getBalance();
-  //   expect(afterBalance).to.not.deep.eq(beforeBalance);
-  //   expect(afterBalance.confirmed).to.be.gt(beforeBalance.confirmed);
-  //   await wallet.lock();
-  // });
+    const { web3 } = await worker.getWeb3();
+    await sendTransaction('geth', addresses[0], web3.utils.toWei('.01', 'ether'), web3, wallet);
+    await sawBlock;
+    await worker.disconnect();
+    await worker.stop();
+    const afterBalance = await wallet.getBalance();
+    expect(afterBalance).to.not.deep.eq(beforeBalance);
+    expect(afterBalance.confirmed).to.be.gt(beforeBalance.confirmed);
+    await wallet.lock();
+  });
 
   it('should have receipts on tx history', async () => {
     const wallet = await getWallet();
@@ -159,7 +161,7 @@ describe('Polygon', function() {
     const wallet = await getWallet();
     const addresses = await wallet.getAddresses();
 
-    const worker = new EthP2pWorker({ chain, network, chainConfig });
+    const worker = new EVMP2pWorker({ chain, network, chainConfig });
     const done = worker.syncDone();
     const sawBlock = new Promise(resolve => worker.events.on('block', resolve));
     await worker.start();
@@ -171,7 +173,7 @@ describe('Polygon', function() {
     await done;
     await worker.stop();
 
-    const dbBlocks = await EthBlockStorage.collection.count({ chain, network });
+    const dbBlocks = await EVMBlockStorage.collection.count({ chain, network });
     expect(dbBlocks).to.be.gt(0);
     await wallet.lock();
   });
